fix: remove dead settings modal, add sign-out to Settings profile section (#776)
The UserAdminPanel's internal User Settings modal was unreachable since
onOpenSettings is always provided, routing to the SettingsPanel instead.
Removes the dead modal block, its state, and all supporting code. Adds a
Sign out button to ProfileSection in SettingsPanel where users will find it.
